fix: 헤더 분리 (#101)
